requirements.txt

streamlit
openai
chromadb
pysqlite3-binary
PyPDF2
protobuf==3.20.*

